refactor(settings_ui): extract supervisor options + self-restart to _supervisor.py
Repoints test patch targets and tools_dev imports to the canonical
ha_mcp.settings_ui._supervisor.<name> site; handlers reach the helpers
via _supervisor.<name> attribute access so one patch site covers all
callers.
